% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2f.R
\name{t2f}
\alias{t2f}
\title{Title: Function to export a dataframe or table to a pdf figure.}
\usage{
t2f(
  df,
  tabname,
  size = 10,
  width = "1in",
  rownames = FALSE,
  scolor = "gray!6",
  strps = NULL,
  pack = NULL,
  digits = 3
)
}
\arguments{
\item{df}{A dataframe or table to be exported.}
}
\description{
Description: A detailed explanation of the function.
}
\examples{
t2f(df)

}
